Update validation.py
i have create a validation.py in the backend which checks the seating arrangement and take the hall as the input from the seating_arrangement.py module i have added two hall allocation for example if the allocated hall is  wrong it will display what is the wrong in that and mainly i check Matrix Processing and Hash Mapping: It processes the hall arrangement as a 2D matrix and maps department codes (bcs, bee, ece) to specific hash values using dictionaries.
Validation with Debugging: Implements validation logic to identify conflicts and invalid entries, with optional debugging to display internal states like department hash matrices.